refactor: put command/mediator logic to Fabron.Mando
