+++ src/Core.Test/Grains/TestCommand.cs
@@ -1,6 +1,6 @@
 using System.Text.Json;
-using Fabron.Contracts;
 using Fabron.Grains;
+using Fabron.Mando;
 
 namespace Fabron.Test.Grains
 {
