+++ src/Core/Contracts/JobMappings.cs
@@ -6,7 +6,7 @@ using Fabron.Grains.BatchJob;
 using Fabron.Grains.CronJob;
 //using Fabron.Grains.CronJob;
 using Fabron.Grains.TransientJob;
-using Fabron.Services;
+using Fabron.Mando;
 
 namespace Fabron.Contracts
 {
