+++ src/Core/Grains/TransientJob/TransientJobGrain.cs
@@ -5,7 +5,7 @@ using Microsoft.Extensions.Logging;
 using Orleans;
 using Orleans.Concurrency;
 using Orleans.Runtime;
-using Fabron.Services;
+using Fabron.Mando;
 
 namespace Fabron.Grains.TransientJob
 {
