+++ src/Core/HostBuilderExtensions.cs
@@ -4,7 +4,7 @@ using Orleans;
 using Orleans.Hosting;
 using Fabron;
 using Fabron.Grains.TransientJob;
-using Fabron.Services;
+using Fabron.Mando;
 
 namespace Microsoft.Extensions.Hosting
 {
