+++ src/Core/IJobManager.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using Fabron.Contracts;
+using Fabron.Mando;
 
 namespace Fabron
 {
