+++ src/Core/JobManager.Batch.cs
@@ -9,6 +9,7 @@ using System.Threading.Tasks;
 using Microsoft.Extensions.Logging;
 using Fabron.Contracts;
 using Fabron.Grains.BatchJob;
+using Fabron.Mando;
 
 namespace Fabron
 {
