+++ src/Core/JobManager.Cron.cs
@@ -11,6 +11,7 @@ using Microsoft.Extensions.Logging;
 using Fabron.Contracts;
 using Fabron.Grains.CronJob;
 using Fabron.Grains.TransientJob;
+using Fabron.Mando;
 
 namespace Fabron
 {
