+++ src/Core/JobManager.Transient.cs
@@ -7,6 +7,7 @@ using System.Threading.Tasks;
 using Microsoft.Extensions.Logging;
 using Fabron.Contracts;
 using Fabron.Grains.TransientJob;
+using Fabron.Mando;
 
 namespace Fabron
 {
