+++ src/Mando/Mediator.cs
@@ -3,7 +3,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.Extensions.Options;
 
-namespace Fabron.Services
+namespace Fabron.Mando
 {
     public interface IMediator
     {
